Implemented C grammar
The entirety of the C grammar, with the exception of the preprocessor, has been implemented. In addition, the lexer can now be constructed to detect newlines as significant tokens, separate from whitespace (only used in the preprocessor). Known bug: the parser's AST output does not create any branches -- it is all folded into one large node.